Simplify Makefile for Contacts Admin having simpler seeds
In https://github.com/alphagov/contacts-admin/pull/963 we simplified
Contacts Admin to not need the elaborate seed set-up (which involved a
chain of applications needing to be running).

This then simplifies the E2E test configuration for this by setting up
contacts-admin to run db:seeds in the regular way.

An additional step was needed in Whitehall to publish organisations to
the Publishing API. This is needed so that an organisation can exist for
the contacts finder