Add a healthcheck for Publishing API
I'm adding this in to resolve an issue where applications will try to
publish to the Publishing API before it is available. In the past we
managed to side step this problem by running the task to seed contacts
admin [1] which presumably took long enough that Publishing API was started.

We have a step that waits for a healthy Publishing API, but since there
isn't actually a healthcheck for Publishing API this is of minimal (if
any) value.

Typically we've set up healthchecks in Dockerfiles, but there isn't one
currently [2] and the trend seems to be to towards removing those [2],
therefore adding one here seems the appropriate place.

[1]: https://github.com/alphagov/publishing-e2e-tests/blob/2325fba0901ff42a6b68ddf44ca837d52f260e5e/Makefile#L62-L63
[2]: https://github.com/alphagov/publishing-api/blob/caac2d99289db453ae5c4bf9dd0517ee3a9f662e/Dockerfile
[3]: https://github.com/alphagov/government-frontend/commit/0e2b735f0d5a4712b96a4a88fdc4828f6ceecbcd