Remove the Preview Design System permission for Travel Advice tests
This is added by default for seed data. We are currently working
on this bit of functionality behind a feature flag and it is
not currently ready for E2E tests and to be shipped.

Once ready we will port over the E2E tests to use the new flow.