Shorten guidance about adding new tests
The paragraph about consulting other teams is less relevant today,
as we no longer have a dedicated team working on this repo.

The paragraph about flakey tests is covered by a later paragraph,
which I've revised to highlight the main point of using "new".

The bullet points no longer made sense with the preamble gone, so
I've rephrased them as a list of instructions.

The comment about #end-to-end tests was out-of-date and I think
it's time to remove this functionality [^1].

[^1]: https://github.com/alphagov/publishing-e2e-tests/pull/475