Remove Slack failure notifications for this repo
I'm not aware anyone is monitoring this channel, especially when
there is no specific team maintaining the repo. The artifical
environment the tests run in is also not relevant for any kind of
transient failure checks - comparing with Smokey, for example.

We still have good visibility of failures as part of the CI checks
on all apps involved in this repo as well as PRs on the repo itself.

Note: the slackSend function comes from the Jenkins Slack plugin [^1],
which is used elsewhere so there's no further cleanup needed here.

[^1]: https://www.jenkins.io/doc/pipeline/steps/slack/