Get ContactAdmin update contact test passing
The update contact test is currently intermittently failing. The failing
screenshot of shows the non-updated contact.

While looking at some of the other tests to see if they had any
similarities i found that the Whitehall update document spec had run
into the issue and fixed it by using the Session#find method. This is
used to identify when the element containing the updated title appears
before attempting to validate it's presence.

Currently, it is unclear why this resolves the issue as the page is not
being dynamically rendered, therefore theoretically
waiting for an element to appear should not work.

Kevin Dew and I have discussed this and decided that we're going to
merge this despite not understanding why the fix works as it's a core
bit of functionality.

If this test begins to flake in future we can mark it as flaky as
outlined here https://github.com/alphagov/publishing-e2e-tests/blob/main/docs/troubleshooting.md#dealing-with-a-flaky-test