Wait for uploaded asset to exist
The test for uploading an attachment in Whitehall was failing when receiving a HTTP 404 Not Found error.

This resulted in the test suite failing frequently, presumably due to a race condition between the 'publish' action and the 'preview' step of this test.

This change follows in the footsteps of #496 and #104.