Adapt tests to work with Miller Columns checkbox
These tests broke following the introduction of a new 'Taxonomy topic tags' page in alphagov/whitehall#6893. The new page design uses the [miller-columns element][] to present the list of tags for users to choose from.

1. Label is not directly clickable
==================================

The tests struggled to locate and check checkboxes using the usual approach:
```
check "Label of my checkbox"
```

Instead, the following exception would be raised:
```
*** Selenium::WebDriver::Error::ElementClickInterceptedError Exception: element click intercepted: Element <label class="govuk-label govuk-checkboxes__label" for="taxonomy_tag_form-taxons-0">...</label> is not clickable at point (298, 1070). Other element would receive the click: <div class="govuk-checkboxes__item">...</div>
  (Session info: headless chrome=106.0.5249.91)
```

It seems like the `<label>`'s parent element was intercepting the click event, meaning the usual `check` helper method couldn't be used. So instead we find the parent and click that.

[miller-columns element]: https://alphagov.github.io/miller-columns-element/

2. Don't uncheck the topic from the topic taxonomy
==================================================

The change above results in slightly different behaviour in the tests.

The previously used `check` method would ensure that a checkbox was checked. If the checkbox was already checked, the `check` method will 'do nothing' – i.e. it won't uncheck the checkbox, but instead leave it checked.

But the new approach (i.e. clicking the label's parent to cause the checkbox to become checked) means that subsequent clicks will uncheck the checkbox again.

This caused this test to fail as when the content is updated it ended up removing the topic that was previously selected. To resolve this I've removed these steps from the journey so that we no longer select a topic when updating content.